Move db helper functions for lobbySocket into 1 file
